Introduce optimisation directory
* move grid search and Bayesian optimisation files to optimisation
  sub-directory

* initialise most members of XGBoostClassifierBayesianOpt in
  template_xgboost.py to keep control about these things in one place
