+++ machine_learning_hep/templates_xgboost.py
@@ -20,6 +20,7 @@ from xgboost import XGBClassifier
 from hyperopt import hp
 
 from machine_learning_hep.optimisation.bayesian_opt import BayesianOpt
+from machine_learning_hep.optimisation.metrics import get_scorers
 
 def xgboost_classifier(model_config): # pylint: disable=W0613
     return XGBClassifier(verbosity=1,
@@ -45,7 +46,7 @@ class XGBoostClassifierBayesianOpt(BayesianOpt):
 
     def yield_model_(self, model_config, space):
         config = self.next_params(space)
-        config["early_stopping_rounds"] = 10
+        config["early_stopping_rounds"] = 50
         # NOTE If that's not really an integer, it will crash!
         if "n_estimators" in config:
             config["n_estimators"] = int(config["n_estimators"])
@@ -62,4 +63,10 @@ class XGBoostClassifierBayesianOpt(BayesianOpt):
 
 
 def xgboost_classifier_bayesian_opt(model_config):
-    return XGBoostClassifierBayesianOpt(model_config, xgboost_classifier_bayesian_space())
+    bayesian_opt = XGBoostClassifierBayesianOpt(model_config, xgboost_classifier_bayesian_space())
+    bayesian_opt.nkfolds = 3
+    bayesian_opt.scoring = get_scorers(["AUC", "Accuracy"])
+    bayesian_opt.scoring_opt = "AUC"
+    bayesian_opt.low_is_better = False
+    bayesian_opt.n_trials = 50
+    return bayesian_opt
